+++ RELEASE_NOTES.md
@@ -1,8 +1,7 @@
 [![Maven Central](https://img.shields.io/static/v1?label=MavenCentral&message=0.0.17-SNAPSHOT&color=blue)](https://search.maven.org/artifact/de.skuzzle.springboot.test/spring-boot-wiremock-parent/0.0.17-SNAPSHOT/jar) [![JavaDoc](https://img.shields.io/static/v1?label=JavaDoc&message=0.0.17-SNAPSHOT&color=orange)](http://www.javadoc.io/doc/de.skuzzle.springboot.test/spring-boot-wiremock-parent/0.0.17-SNAPSHOT)
 
-### Bug Fixes
-* [#2](https://github.com/skuzzle/spring-boot-wiremock/issues/2) Eventually fix locating the keystore problem 
-* [#11](https://github.com/skuzzle/spring-boot-wiremock/issues/11) Fix problem with properties not being injected into `ConfigurationProperties`
+### Features:
+* Add `@Response.withJsonBody` which allows to specify a json response body and automatically sets the `Content-Type` to `application/json` 
 
 ### Maven Central coordinates for this release